tess_page/requirements.txt
pytesseract==0.3.0
lxml==4.2.1
image==1.5.27
Pillow==8.2.0
